requirements.txt
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
drf-spectacular==0.28.0
isort==7.0.0
mysqlclient==2.2.7
pre_commit==4.3.0
python-dotenv==1.1.1
requests==2.32.5
str2bool==1.1
uWSGI==2.0.31
